Modifies sanitizeCallback integration test to use AdminTestCase.
